backend/requirements-prod.txt
fastapi
uvicorn[standard]
opencv-python
numpy
insightface
python-multipart
onnxruntime
gunicorn
python-dotenv
pydantic-settings
python-json-logger
